Demo complete cycle of a GET request with Redux Saga
Click on "Fetch Pandas". This fires off a GET request to reddit/aww (hey, I needed some kind of endpoint). The user is shown an appropriate message depending on whether the request was successful or not.

Also:
- Set up tests to be Redux friendly
- Connect the Home component to the store
- Add support for [Redux dev tools](https://github.com/zalmoxisus/redux-devtools-extension)
- Test the undecorated version of the components:
In App and Home, we are exporting both the unconnected and the connected versions of the component. The connected version is the default export, and it is used in the app itself (i.e. in the Routes). The undecorated version is imported by the tests, since we don't really care about the Redux store or the react-redux wrapper when testing these components.

Reference: http://stackoverflow.com/questions/41210052/react-redux-testing-w-enzyme#41210100