Replace Material UI DatePicker with homemade Calendar
Add utility to find the date that should appear in the upper left position of a calendar
Add a utility to find the first day in each week in a month
Calendar defaults to displaying current month
Highlight the selected date
Fix leftPane styling. Error messages now stack below the other left pane content.
Add component for DayNames
Add DateDisplay component. This is responsible for displaying the selected date above the calendar. It includes the past/future navigation arrows, but I didn't wire in any functionality because it's not clear what they should do.
Clicking date updates the date in UI. It does not, however, update the Agenda to show the meetings for the selected date.

I put the most important calendar sizing properties into a single config file. So if, for instance, we need to adjust the spacing between the dates while also resizing the calendar itself, we can do it all in one spot. Hopefully this makes future styling changes less insanity-inducing.