Improve Meeting Cancellation flow
Clicking on a meeting opens edit form, populated with the meeting info.
The meeting form has two modes: meeting creation and meeting editing. The mode switch is triggered by `state.app.isCreatingMeeting`.
Clicking on 'Delete' opens the cancellation dialog.
Add action to reset the UI in one fell stroke, which is especially useful for clearing all messages.
Give a visual indication of disabled button. For instance, the 'Bookit' button is disabled until the form is valid. Note that the 'Save' button, which is seen in the Meeting Edit mode, is always disabled. This is because we have not yet implemented Meeting Editing.