Flatten meetings data
The meetings data as received from the server is deeply nested, which has caused our components to perform all sorts of weird contortions. The reducer logic was overly complex, not to mention the weird mappings and reductions we had going on in several components.

This commit flattens the meeting data. Hopefully this will make the application state much easier to work with. The logic for retrieving or updating a given item is now fairly simple and consistent. For instance, given a meeting's ID, we can directly look it up and figure out which room it's in a couple simple steps, without having to dig through other objects to do so. No more `roomMeeting.room.meeting.room`!

Here's how the Actions, Containers, and Dumb Components work together under this scheme:

== Actions ==
The application state is flattened in the Action that receives the fetched meetings.

== Containers ==
Also known as "connected components". This is where the Redux store enters the app. `mapStateToProps` is used to modify parts of the state tree in ways specifically catered to the components it contains. Things like `isOwnedByUser` are defined at this level. This is also the place to filter the list of meetings, by, for instance, the date of the meeting.

== Dumb Components ==
There is little, if any manipulation of data in the components themselves. They are responsible for simply rendering the data they are given. Dumb Components are concerned with things like semantic markup and style.