Restructure the Dashboard component
The `Dashboard` component was handling a lot of different things at once -- probably too much. This commit creates `Header` and `InfoPanel` components, which are children of `Dashboard`.

This `InfoPanel` is not only badly named, it also serves as a container for the `Calendar`, the `ReservationList`, and the various forms that appear on the left.