Admin user can remove a user
This is a rough first pass at this functionality. Currently, clicking "Remove" removes the user immediately, with no intermediate confirmation dialog.

Also, this only removes the user from the Redux state. It does not make a request to the server to remove the user, since the server-side of this feature is out of current scope.