Admin remove user: Add confirmation dialog. Handle complete request cycle.
When the admin user clicks remove, they see a confirmation dialog asking if they are sure they'd like to remove the user. If yes, the user is removed from the Redux store and the user sees a nice message. (The api call to the server can be dropped into the `users` saga, once the endpoint is ready.) If the user clicks "No", the dialog is closed.