Fetch and display list of external users
When the Admin screen loads, the app fetches the list of Wipro users
from bookit-server and shows them on the screen. Click on the "Wipro"
tab to see this list.

This commit implements only the fetching of external users, which is why
the "Designit" list is blank.

This commits also fully implements recently added users. We initially thought this might be tough
to do, but it turns out the MS api provides a "createdDateTime" field
that we can translate into a "dateAdded" property. With this commit, the
"Recently added users" list functions as expected with real data.

This commit makes it clear that removing users is not fully functional by showing the user a message. This feature was previously client-side only, by which I mean that it removes the user from the Redux store, but it does not make an api call and therefore does not actually remove the user from the datastore. This commit seeks to make it really clear exactly where we're at in terms of implementing this feature.

Known gotcha: The external team, Wipro, is defined in the code. In fact,
the internal team is as well. This obviously does not lend itself well
to extending Bookit to other use cases. It's also brittle, since we're
relying on the client and the server to agree to a convention. I chose
not to address this issue in this bit of work, since doing it right will
involve a larger discussion around configuration across client and
server.