Refactor forms
This commit refactors the two forms in our app: the meeting form and the add user form. It is based on the spike we did earlier this week, which can be found here: https://github.com/BillyZac/spike-redux-forms

Validations are now much cleaner, and protected by tests.

We created a stub prop to allow for implementing "QuickBooking". See `src/containers/MeetingForm' lines 94 and 95.

The forms are defined in one file each. (Previously each form was spread across a "container" and a "component".)

Made DateTimePicker a controlled set of Kronos components to make behaviour management easier.

Adds tests for validations and basic tests for meeting form. We could benefit from writing tests to validate click behavior, etc, but we stopped short when it felt like we were testing the behavior of Kronos itself.

The current tests handle the basic case of showing the correct form title depending on whether you're editing or creating a booking. We could stand to test the MeetingForm more deeply, but at this point we're not sure how to use Enzyme to test a component generated by redux-form.